LayoutTests/js/exception-properties-expected.txt

Test for correct properties on Error objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS enumerableProperties(error).sort() is ["column", "line", "sourceURL"]
PASS enumerableProperties(nativeError).sort() is ["column", "line", "sourceURL"]
PASS Object.getPrototypeOf(nativeError).name is "RangeError"
PASS Object.getPrototypeOf(nativeError).message is ""
PASS successfullyParsed is true

TEST COMPLETE
